done with Authorization

+++ backend/routes/posts.js
@@ -41,8 +41,10 @@ router.post(
     const post = new Post({
       title: req.body.title,
       content: req.body.content,
-      imagePath: url + "/images/" + req.file.filename
+      imagePath: url + "/images/" + req.file.filename,
+      creator:req.userData.userId
     });
+    console.log(req.userData);
     post.save().then(createdPost => {
       res.status(201).json({
         message: "Post added successfully",
@@ -68,11 +70,19 @@ router.put(
       _id: req.body.id,
       title: req.body.title,
       content: req.body.content,
-      imagePath: imagePath
+      imagePath: imagePath,
+      creator:req.userData.userId
     });
     console.log(post);
-    Post.updateOne({ _id: req.params.id }, post).then(result => {
+    Post.updateOne({ _id: req.params.id, creator: req.userData.userId}, post).then(result => {
+      console.log(result);
+      if(result.nModified > 0){
       res.status(200).json({ message: "Update successful!" });
+      }else{
+        res.status(401).json({
+          message:"Un Authorized Access!"
+        })
+      }
     });
   }
 );
@@ -95,7 +105,7 @@ router.get("", (req, res, next) => {
    postQuery
    .then(documents => {
      fetchedPosts=documents;
-     return Post.count();
+     return Post.countDocuments();
     // res.status(200).json({
     //   message: "Posts fetched successfully!",
     //   posts: documents
@@ -120,10 +130,16 @@ router.get("/:id", (req, res, next) => {
 });
 
 router.delete("/:id",checkAuth, (req, res, next) => {
-  Post.deleteOne({ _id: req.params.id }).then(result => {
+  Post.deleteOne({ _id: req.params.id, creator: req.userData.userId }).then(result => {
     console.log(result);
-    res.status(200).json({ message: "Post deleted!" });
-  });
+if(result.n> 0){
+      res.status(200).json({ message: "deletion successful!" });
+      }else{
+        res.status(401).json({
+          message:"Un Authorized Access!"
+        })
+      }
+    });
 });
 
 module.exports = router;
